+++ src/endpoints/file.rs
@@ -12,7 +12,7 @@ use crate::endpoints::{Endpoint, EndpointConfig};
 use crate::notifications::{Key, ValidatedNotification};
 use crate::{Error, LIB_LOG_TARGET};
 use async_trait::async_trait;
-use log::{info, warn};
+use tracing::{info, warn};
 use serde::Deserialize;
 use std::any::Any;
 use std::collections::{HashMap, HashSet};