+++ src/notifications.rs
@@ -75,15 +75,13 @@ impl Notification {
     pub fn to_json(&self) -> Result<String, Error> {
         Ok(serde_json::to_string(self)?)
     }
-
-    #[allow(dead_code)]
+    
     /// Compare provided notification name ['Key'] to this notification.
     pub(crate) fn validate(&self, hash_key: &Key) -> bool {
         let new_key = self.message.create_key(hash_key);
         self.key == new_key.to_hex()
     }
-
-    #[allow(dead_code)]
+    
     /// Compare provided set of  notification name ['Key']s to this notification.
     pub(crate) fn validate_set(&self, hash_keys: &HashSet<Key>) -> bool {
         for hash_key in hash_keys {